Add files via upload
The project is based on general file handling concept. The project focuses on IPL TICKET PAYMENT .The project links a external text file to store and manipulate the data's. The project also secures the data by providing password ("keerthu123") so that the admins only can manipulate or design the billing amount. The payment process and due list is made as public and user friendly so that people can access it conveniently.